Added SubmissionForProcessing, submissions in queue in /submissions page and background job for automatic resubmission (#774)
Closes: 
https://github.com/SoftUni-Internal/exam-systems-issues/issues/794 
and
https://github.com/SoftUni-Internal/exam-systems-issues/issues/698

This pull request implements the following changes:

1. Usage of SubmissionsForProcessing table - Any new submission is saved
in both the Submissions and SubmissionsForProcessing tables. The
SubmissionsForProcessing table acts as a queue in the database. The
following scenarios are possible for entries in
SubmissionsForProcessing:
  - processing=false; processed=false - The submission was just added
- processing=true; processed=false - The submission was sent to the
message queue (status is "In Queue")
- processing=false; processed=true - The execution result was received
for the submission
- The submission is deleted from SubmissionsForProcessing after
successfully processing the execution result

2. Pending submissions background job - A Hangfire job runs every 3
minutes to check the SubmissionsForProcessing table for pending
submissions (processing=false and processed=false) that are older than 1
minute based on CreatedOn. This handles cases where there was an error
publishing the submission to the queue (e.g. queue was down). The job
publishes these submissions to RabbitMQ.

The background job is added via `GET` to
https://admin.dev.alpha.judge.softuni.org/BackgroundJobs/AddRecurringEnqueuePendingSubmissionsForProcessingJob
endpoint and can be monitored at the Hangfire dashboard
https://admin.dev.alpha.judge.softuni.org/hangfire

3. FE changes


![image](https://github.com/SoftUni-Internal/OpenJudgeSystem/assets/19411274/1fc83994-00a4-4d73-94f5-add23ccc2f8d)


The following component should be visible to admins. Key points:
- Submissions awaiting execution - the count of submissions that are
**procecessing (processing=true; processed=false)** or **pending
(processing=false; processed=false)**
- All / In Queue/ Pending links for showing submissions in different
states

---------

Co-authored-by: Petar Neshkov <petar.neshkov@softuni.bg>
Co-authored-by: Simona Simeonova <simona.simeonova@softuni.bg>
Co-authored-by: Georgi Georgiev <gogo4ds@gmail.com>
Co-authored-by: Simona Simeonova <simona.simeonova@hotmail.com>