SpeedUp callbacks and send email when redis is down. (#803)
Closes
https://github.com/SoftUni-Internal/exam-systems-issues/issues/883
Closes
https://github.com/SoftUni-Internal/exam-systems-issues/issues/880

Connection multiplex was registered in the app so it can be used to
check weather the redis cache is connected or not.
Then a method is called which checks if the redis is connected. If so it
continues with the action callback, if it is not connected then sends an
email and continues with fallback function.