Add default strategy for problem (#864)
Closes
https://github.com/SoftUni-Internal/exam-systems-issues/issues/928

Added new column in Problem table.
On update and create Views for problem, new dropdown list is added which
will hold the values of the selected submission types.
The Selected index of the strategy represents the default submission
type for the problem.
On code editor page , in the dropdown list for submission types. the
default for the problem will be selected automatically (TimeLimit,
MemoryLimit.. etc will be updated).