Contest card fix to allow participant to compete after Contest time has ended (#874)
Closes:
https://github.com/SoftUni-Internal/exam-systems-issues/issues/927


**Summary of the changes made**:
1. Moving ContestsActivityService, UserProviderService and
ParticipantsCommonDataService and their interfaces to Services/Common
2. Updating logic and adding additional methods to
ContestActivityService
3. Removing properties - CanBeCompeted, CanBePracticed and IsActive from
Contest model and replacing usage with the activity service
4. Fixing Contest card's Compete/Practice buttons so that they are
always active for admins and also to stay enabled/active if current used
is a participant in online exam and has participationTime left but
contest end time has timed out.
*In order to keep the contest card active for participants in online
exams the compete button doesn't deactivates after the time on the
contest card is timed out. Those changes were introduced recently but
are no longer serving correct purpose and lead to difficulties. User see
the buttons as active until he refreshes the page but if clicks on
compete and both the contest is expired and he doesn't have any
participation time left then an error message for expired contest is
shown.