Added a validation to ensure the the user will be redirected only when not logged in.
Added a validation that checks the status of the query that fetches the resource data, and if the error status is 401, then the user will be redirected to the login page, otherwise a simple message will be displayed. Reduced the time before the user is redirected to 3s instead of 5s.