The buttons should not be disabled if the user is invalidated.
Previously the contest's buttons for compete / practice were disabled when an administrator changed the contest's password, because we were checking whether the user is currently active, which means that he should not be invalidated. But when the passwords are changed, the user is invalidated. The buttons should not be active only when the user's compete / participation time has ran out. This way if an admin changes the password while the user is participating / competing in the contest, they will be required to provide the new password and the buttons will not be disabled.