Added a dynamic '/:slug' path to the 'Contest Details' and 'All Contests' pages.
After changing the contest's details page to use a slug, two paths match: '/contests/:contestId/:slug' and '/contests/:contestId/:participationType', that is why it was decided to add a static path '/details' to the contest's details page. This way the Router will be able to differentiate between the two paths, since React Router doesn't support dynamic path matching with a regex.