Fixed the issue regarding a weird link showing up on the 'Submission Details' page.
The issue was caused due to not having any styles applied on the link, as well as the data that the link displays being undefined ( the data displayed is the contest's name ). The data is fetched from the redux store, whilst not being set. The solution is, before the  submission's details button is clicked, set the needed data in the redux store.