Fixed a bug in the IsEligibleForRetest method.
Since the tests are no longer fetched with the submission itself, but with the test runs, if there are not test runs, no tests will be fetched ( even if they exist ). This is why a separate query must be made to the database, when the Tests' count is 0, this may not seem optimal, but the amount of times a user will be retesting a submission is far less often than the amount of times they will be checking their submission's details when the test runs exist ( the submission is not eligable for a retest ).